Änderungen von Eric

+++ Client/controller/imageupdater.h
@@ -48,7 +48,7 @@ private slots:
     *  	@param[in] _posData			Neue Position
     */
     void NewPosition(PosData*);
-    void drawNewLine(int x1, int y1, int x2, int y2);
+    void drawNewLine(double x1, double y1, double x2, double y2);
 
 private:
